# Backned/FinSenseAPI/FinSenseAPI/DTOs/Auth/AuthResponseDto.cs
using System;

namespace FinSenseAPI.DTOs.Auth;

public class AuthResponseDto
{
    public string Token { get; set; }
    public string FullName { get; set; }
    public string Email { get; set; }
    public DateTime ExpiresAt { get; set; }
    public string RefreshToken { get; set; }

}