# Backned/FinSenseAPI/FinSenseAPI/DTOs/Auth/RegisterRequestDto.cs
using System;

namespace FinSenseAPI.DTOs.Auth;

public class RegisterRequestDto
{
    public string FullName { get; set; }
    public string Email { get; set; }
    public string Password { get; set; }

}